+++ src/App.js
@@ -1,12 +1,10 @@
 import logo from "./logo.svg";
 import "./App.css";
 import React from "react";
-import Navbar from "./animations/navbar";
-import Contact from "./Pages/contact";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Developers from "./Pages/developers";
-import Home from "./components/Home";
-import Team from "./Pages/team";
+import Home from "./Pages/home/Home";
+import Team from "./Pages/team/team";
 
 function App() {
   return (
