+++ src/Pages/home/Home.css
@@ -1,0 +1,262 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400&display=swap");
+.home {
+  position: relative;
+  height: 100vh;
+  overflow: hidden;
+  background: url(../../images/b8.jpg);
+  background-attachment: fixed;
+  background-size: cover;
+  width: 100vw;
+  background-position-x: center;
+  font-family: "Roboto Slab", serif;
+}
+
+.home-content {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  padding: 50px 70px 0;
+  height: 100%;
+}
+
+.home-text {
+  width: 100%;
+  padding-bottom: 150px;
+  font-size: 150px;
+  line-height: 46px;
+  color: #fff;
+  letter-spacing: 10px;
+  text-align: center;
+  margin-top: 15%;
+  margin-left: 20%;
+  margin-right: 20%;
+}
+
+@media screen and (max-width: 500px) {
+  .home-content,
+  .home-img {
+    width: 100%;
+    height: 100%;
+    padding: 0;
+    padding-bottom: 50px;
+  }
+}
+
+.home-text-div {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+}
+.home-text-div span {
+  position: absolute;
+}
+.home-text-div span:nth-child(1) {
+  top: 20%;
+  left: 0%;
+}
+.home-text-div span:nth-child(1)::before {
+  left: 45%;
+}
+.home-text-div span:nth-child(2) {
+  top: 0%;
+  left: 20%;
+}
+.home-text-div span:nth-child(2)::before {
+  left: 25%;
+}
+.home-text-div span:nth-child(3) {
+  top: 20%;
+  left: 40%;
+  transform: rotate(20deg);
+}
+.home-text-div span:nth-child(3)::before {
+  left: -40%;
+  height: 750%;
+  transform: rotate(-20deg);
+}
+.home-text-div span:nth-child(4) {
+  top: 0%;
+  left: 60%;
+}
+.home-text-div span:nth-child(5) {
+  top: 20%;
+  left: 74%;
+  transform: rotate(-20deg);
+}
+.home-text-div span:nth-child(5)::before {
+  left: 115%;
+  height: 740%;
+  transform: rotate(20deg);
+}
+.home-text-div span:nth-child(6) {
+  top: 0%;
+  left: 95%;
+}
+
+.home-text-div span::before {
+  content: "";
+  position: absolute;
+  background: #fff;
+  opacity: 0.7;
+  width: 5px;
+  height: 700%;
+  justify-content: center;
+  top: -750%;
+  left: 40%;
+}
+
+.footer {
+  position: absolute;
+  bottom: 0;
+  height: 70px;
+  color: #fff;
+  justify-content: center;
+  align-items: center;
+  background: #000;
+  width: 100%;
+}
+.footer p {
+  width: 100%;
+  height: 100%;
+  justify-content: center;
+  text-align: center;
+  margin-top: 20px;
+}
+
+.home-text-div span,
+.home-text-div span::before {
+  animation: animated 12.14s linear infinite;
+}
+
+@keyframes animated {
+  0%,
+  18%,
+  20%,
+  50.1%,
+  60%,
+  65.1%,
+  80%,
+  90.1%,
+  92% {
+    color: #ffffff;
+    text-shadow: none;
+  }
+  30%,
+  33%,
+  60.1%,
+  65%,
+  80.1%,
+  90%,
+  92.1%,
+  100% {
+    color: #fff;
+    text-shadow: 0 0 10px #03bcf4, 0 0 20px #03bcf4, 0 0 40px #03bcf4,
+      0 0 80px #03bcf4, 0 0 160px #03bcf4;
+  }
+}
+
+/* Login */
+
+.login {
+  position: fixed;
+  top: 20px;
+  right: 20px;
+  font-size: 24px;
+  transition: 0.2s;
+  color: #fff;
+  overflow: hidden;
+  border: 2px solid #2196f3;
+  background: #000;
+  padding: 10px;
+  padding-left: 20px;
+  padding-right: 20px;
+  border-radius: 10px;
+}
+.login:hover {
+  background: #2196f3;
+  box-shadow: 0 0 10px #2196f3, 0 0 40px #2196f3, 0 0 80px #2196f3;
+  transition-delay: 0.25s;
+  cursor: pointer;
+}
+.login span {
+  position: absolute;
+  display: block;
+}
+
+@media screen and (max-width: 600px) {
+  .home-text-div {
+    font-size: 70px;
+  }
+  .home-text-div span::before {
+    content: "";
+    position: absolute;
+    background: #fff;
+    opacity: 0.7;
+    width: 3px;
+    height: 700%;
+    justify-content: center;
+    top: -750%;
+    left: 40%;
+  }
+  .home-text {
+    width: 100%;
+    margin-left: 10%;
+    margin-right: 20%;
+  }
+  .home-text-div span:nth-child(1) {
+    top: 45%;
+    left: 0%;
+  }
+  .home-text-div span:nth-child(1)::before {
+    left: 40%;
+    height: 751%;
+  }
+  .home-text-div span:nth-child(2) {
+    top: 30%;
+    left: 20%;
+  }
+  .home-text-div span:nth-child(2)::before {
+    left: 20%;
+    height: 750%;
+  }
+  .home-text-div span:nth-child(3) {
+    top: 45%;
+    left: 40%;
+    transform: rotate(20deg);
+  }
+  .home-text-div span:nth-child(3)::before {
+    left: -100%;
+    height: 775%;
+    transform: rotate(-20deg);
+  }
+  .home-text-div span:nth-child(4) {
+    top: 30%;
+    left: 60%;
+  }
+  .home-text-div span:nth-child(4)::before {
+    left: 25%;
+    height: 750%;
+  }
+  .home-text-div span:nth-child(5) {
+    top: 45%;
+    left: 74%;
+    transform: rotate(-20deg);
+  }
+  .home-text-div span:nth-child(5)::before {
+    left: 155%;
+    height: 775%;
+    transform: rotate(20deg);
+  }
+  .home-text-div span:nth-child(6) {
+    top: 30%;
+    left: 95%;
+  }
+  .home-text-div span:nth-child(6)::before {
+    left: 30%;
+    height: 750%;
+  }
+  .login {
+    right: 30%;
+    top: 50%;
+  }
+}
