--- src/animations/developers.css
@@ -1,50 +1,0 @@
-
-.developers
-{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-    background: #050505;
-}
-.cube
-{
-    position: relative;
-    width: 300px;
-    height: 300px;
-    transform-style: preserve-3d;
-    transform:rotateX(-30deg);
-    animation: animate 8s linear infinite;
-}
-
-@keyframes animate
-{
-    0%
-    {
-        transform:rotateX(-30deg) rotateY(0deg);
-    }
-    100%
-    {
-        transform:rotateX(-30deg) rotateY(360deg);
-    }
-}
-.cube div
-{
-    position: absolute;
-    top:0;
-    left:0;
-    width: 100%;
-    height: 100%;
-    transform-style: preserve-3d;
-}
-
-.cube div span
-{
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(#151515,#00ec00);
-    transform:rotateY(calc(90deg*var(--i))) translateZ(150px)
-}
