--- src/animations/developers.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-
-import './developers.css'
-export default function developers() {
-    return (
-        <div className='developers'>
-            <div className="cube">
-                <div className="">
-                    <span className="--i:0">hello</span>
-                    <span className="--i:1"></span>
-                    <span className="--i:2"></span>
-                    <span className="--i:3"></span>
-                </div>
-            </div>
-        </div>
-    )
-}
