--- src/animations/navbar.css
@@ -1,128 +1,0 @@
-.navbar
-{
-    background: #582949;
-    justify-content: center;
-    display: flex;
-    width: 100vw;
-}
-
-.navigation
-{
-    position: relative;
-    height: 80px;
-    width: 100%;
-    background: black;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 10px;
-}
-.navigation ul
-{
-    display: flex;
-    list-style-type: none;
-}
-.list
-{
-    position: relative;
-    width: 80px;
-    height:80px;
-    background: none;
-    border:none;
-    outline: none;
-}
-.list .link
-{
-    display: flex;
-    position: relative;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    flex-direction: column;
-    font-weight: 500;
-    text-align: center;
-    z-index: 1;
-}
-.list .link .icon
-{
-    position: relative;
-    display: block;
-    line-height: 75px;
-    font-size: 1.5em;
-    text-align: center;
-    transition: 0.5s;
-    color:#FFDEB0;
-}
-.list.active .link .icon 
-{
-    transform:translateY(50px);
-    color: #582958
-}
-.list .link .text
-{
-    position: absolute;
-    font-weight: 700;
-    letter-spacing: 0.05em;
-    transition: 0.5s;
-    opacity: 0;
-    width:80px;
-    color:#FFD6F1
-}
-
-.list.active .link .text
-{
-    opacity: 1;
-    transform:translateY(-16px)
-
-}
-
-.link
-{
-    text-decoration: none;
-}
-
-.indicator
-{
-    position: absolute;
-    width: 80px;
-    height: 80px;
-    top:60%;
-    background: #FFD8B9;
-    border-radius: 50%;
-    border:6px solid black;
-    transition: 0.5s;
-}
-
-.list:nth-child(1).active ~ .indicator
-{
-    transform: translateX(calc(80px*0));
-}
-.list:nth-child(2).active ~ .indicator
-{
-    transform: translateX(calc(80px*1));
-}
-.list:nth-child(3).active ~ .indicator
-{
-    transform: translateX(calc(80px*2));
-}
-
-.list:nth-child(4).active ~ .indicator
-{
-    transform: translateX(calc(80px*3));
-}
-
-.list:nth-child(5).active ~ .indicator
-{
-    transform: translateX(calc(80px*4));
-}
-.list:nth-child(6).active ~ .indicator
-{
-    transform: translateX(calc(80px*5));
-}
-
-.list:nth-child(7).active ~ .indicator
-{
-    transform: translateX(calc(80px*6));
-}
-
-
