--- src/components/Home.css
@@ -1,367 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400&display=swap");
-.home {
-  position: relative;
-  height: 100vh;
-  overflow: hidden;
-  background: url(../images/b8.jpg);
-  background-attachment: fixed;
-  background-size: cover;
-  width: 100vw;
-  background-position-x: center;
-  font-family: "Roboto Slab", serif;
-}
-
-.home-content {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  padding: 50px 70px 0;
-  height: 100%;
-}
-
-.home-text {
-  width: 100%;
-  padding-bottom: 150px;
-  font-size: 150px;
-  line-height: 46px;
-  color: #fff;
-  letter-spacing: 10px;
-  text-align: center;
-  margin-top: 15%;
-  margin-left: 20%;
-  margin-right: 20%;
-}
-
-.home-links {
-  position: relative;
-  height: 100%;
-  width: 100;
-  margin-top: -50px;
-  display: flex;
-  flex-direction: column;
-}
-.home-2,
-.home-1 {
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  display: flex;
-  color: #fff;
-}
-.up {
-  margin-top: -60px;
-}
-
-.redirect-left,
-.redirect-right {
-  height: 150px;
-  align-items: center;
-  justify-content: center;
-  padding-top: 60px;
-  font-size: 25px;
-  position: absolute;
-}
-.line-left,
-.line-right {
-  position: absolute;
-  height: 2px;
-  width: 50px;
-  background-color: #fff;
-}
-
-.home-1 .line-left {
-  margin-left: -150px;
-}
-.home-1 .line-right {
-  margin-right: 150px;
-}
-.home-2 .line-left {
-  margin-left: -300px;
-}
-.home-2 .line-right {
-  margin-left: 345px;
-}
-
-.home-1 .redirect-left {
-  margin-left: -200px;
-}
-
-.diamond1 {
-  height: 150px;
-  width: 130px;
-  clip-path: polygon(50% 0%, 100% 50%, 50% 100%, 0% 50%);
-  justify-content: center;
-  align-items: center;
-  display: flex;
-  background: #fff;
-  margin-left: 20px;
-}
-.diamond1 i {
-  font-size: 50px;
-  margin-top: 20px;
-  width: 100%;
-  height: 100%;
-  color: #fff;
-  text-align: center;
-}
-
-.diamond1:hover {
-  border: 40px solid rgb(127, 80, 255);
-  background: rgb(127, 80, 255);
-  cursor: pointer;
-}
-
-.home-img {
-  position: relative;
-  width: 400px;
-  background: radial-gradient(400px, #fff, transparent 50%);
-  margin-top: -50px;
-}
-.home-img img {
-  width: 50%;
-}
-
-.laptop {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-.home-icons {
-  color: #fff;
-  font-size: 40px;
-  height: 100%;
-  z-index: 2;
-  animation: rotation 60s linear infinite;
-}
-
-.home-icons i {
-  position: absolute;
-  cursor: pointer;
-  z-index: 2;
-  border-radius: 50%;
-  animation: rotation2 60s linear infinite;
-}
-
-@keyframes rotation {
-  100% {
-    transform: rotate(360deg);
-  }
-}
-@keyframes rotation2 {
-  100% {
-    transform: rotate(-360deg);
-  }
-}
-.home-icons i:nth-child(1) {
-  top: 22%;
-  left: 50%;
-}
-.home-icons i:nth-child(2) {
-  top: 35%;
-  right: 10%;
-}
-.home-icons i:nth-child(3) {
-  top: 60%;
-  right: 8%;
-}
-.home-icons i:nth-child(4) {
-  top: 75%;
-  left: 50%;
-}
-.home-icons i:nth-child(5) {
-  top: 60%;
-  left: 8%;
-}
-
-.home-icons i:nth-child(6) {
-  top: 35%;
-  left: 10%;
-}
-
-@media screen and (max-width: 500px) {
-  .home-content,
-  .home-img {
-    width: 100%;
-    height: 100%;
-    padding: 0;
-    padding-bottom: 50px;
-  }
-  .home-text-div {
-    font-size: 70px;
-  }
-}
-
-.home-text-div {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-}
-.home-text-div span {
-  position: absolute;
-}
-.home-text-div span:nth-child(1) {
-  top: 20%;
-  left: 0%;
-}
-.home-text-div span:nth-child(1)::before {
-  left: 45%;
-}
-.home-text-div span:nth-child(2) {
-  top: 0%;
-  left: 20%;
-}
-.home-text-div span:nth-child(2)::before {
-  left: 25%;
-}
-.home-text-div span:nth-child(3) {
-  top: 20%;
-  left: 40%;
-  transform: rotate(20deg);
-}
-.home-text-div span:nth-child(3)::before {
-  left: -40%;
-  height: 750%;
-  transform: rotate(-20deg);
-}
-.home-text-div span:nth-child(4) {
-  top: 0%;
-  left: 60%;
-}
-.home-text-div span:nth-child(5) {
-  top: 20%;
-  left: 74%;
-  transform: rotate(-20deg);
-}
-.home-text-div span:nth-child(5)::before {
-  left: 115%;
-  height: 740%;
-  transform: rotate(20deg);
-}
-.home-text-div span:nth-child(6) {
-  top: 0%;
-  left: 95%;
-}
-
-.home-text-div span::before {
-  content: "";
-  position: absolute;
-  background: #fff;
-  opacity: 0.7;
-  width: 5px;
-  height: 700%;
-  justify-content: center;
-  top: -750%;
-  left: 40%;
-}
-
-@media screen and (max-width: 600px) {
-  .home-text {
-    width: 100%;
-    margin-left: 1%;
-    margin-right: 10%;
-  }
-  .home-text-div span:nth-child(1) {
-    top: 45%;
-    left: 0%;
-  }
-  .home-text-div span:nth-child(1)::before {
-    left: 40%;
-    height: 751%;
-  }
-  .home-text-div span:nth-child(2) {
-    top: 30%;
-    left: 20%;
-  }
-  .home-text-div span:nth-child(2)::before {
-    left: 20%;
-    height: 750%;
-  }
-  .home-text-div span:nth-child(3) {
-    top: 45%;
-    left: 40%;
-    transform: rotate(20deg);
-  }
-  .home-text-div span:nth-child(3)::before {
-    left: -100%;
-    height: 775%;
-    transform: rotate(-20deg);
-  }
-  .home-text-div span:nth-child(4) {
-    top: 30%;
-    left: 60%;
-  }
-  .home-text-div span:nth-child(4)::before {
-    left: 25%;
-    height: 750%;
-  }
-  .home-text-div span:nth-child(5) {
-    top: 45%;
-    left: 74%;
-    transform: rotate(-20deg);
-  }
-  .home-text-div span:nth-child(5)::before {
-    left: 155%;
-    height: 775%;
-    transform: rotate(20deg);
-  }
-  .home-text-div span:nth-child(6) {
-    top: 30%;
-    left: 95%;
-  }
-  .home-text-div span:nth-child(6)::before {
-    left: 30%;
-    height: 750%;
-  }
-}
-
-.footer {
-  position: absolute;
-  bottom: 0;
-  height: 70px;
-  color: #fff;
-  justify-content: center;
-  align-items: center;
-  background: #000;
-  width: 100%;
-}
-.footer p {
-  width: 100%;
-  height: 100%;
-  justify-content: center;
-  text-align: center;
-  margin-top: 20px;
-}
-
-.home-text-div span,
-.home-text-div span::before {
-  animation: animated 12.14s linear infinite;
-}
-
-@keyframes animated {
-  0%,
-  18%,
-  20%,
-  50.1%,
-  60%,
-  65.1%,
-  80%,
-  90.1%,
-  92% {
-    color: #ffffff;
-    text-shadow: none;
-  }
-  30%,
-  33%,
-  60.1%,
-  65%,
-  80.1%,
-  90%,
-  92.1%,
-  100% {
-    color: #fff;
-    text-shadow: 0 0 10px #03bcf4, 0 0 20px #03bcf4, 0 0 40px #03bcf4,
-      0 0 80px #03bcf4, 0 0 160px #03bcf4;
-  }
-}
